Fixed wrong indices used when coloring points

+++ PointCloudAlignement/PointCloudAlignment/include/normal_computation.h
@@ -13,6 +13,6 @@ public:
     void computeNormalCloud(PointNormalKCloud::Ptr cloud_in, KdTreeFlannK::Ptr kdTree_in);
 
 private:
-    float estimateKForPoint(int p_id, PointNormalKCloud::Ptr cloud_in, KdTreeFlannK::Ptr kdTree);
+    int estimateKForPoint(int p_id, PointNormalKCloud::Ptr cloud_in, KdTreeFlannK::Ptr kdTree);
     float computeCurvature(PointNormalKCloud::Ptr cloud, boost::shared_ptr<vector<int>> indices, vector<float> sqrd_distances, PointNormalK p);
 };
